+++ src/http/http2_callbacks.h
@@ -4,6 +4,7 @@
 
 #include "ccf/ds/logger.h"
 #include "ccf/http_configuration.h"
+#include "http/http_exceptions.h"
 #include "http2_types.h"
 #include "http2_utils.h"
